editoast: extract all the STDCM log logic into a function
